pyproject.toml
[project]
name = "fastapi-backend-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.0.1",
    "bson>=0.5.10",
    "fastapi>=0.134.0",
    "passlib==1.7.4",
    "pymongo[srv]>=4.16.0",
    "python-dotenv==1.2.1",
    "python-jose==3.5.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
